moved saveState and loadState to their own util file
